move clear_data function from builder to util module
https://github.com/googlei18n/glyphsLib/pull/84#pullrequestreview-12091599
