+++ Lib/glyphsLib/builder.py
@@ -20,9 +20,10 @@ import logging
 import re
 
 from glyphsLib.anchors import propagate_font_anchors
+from glyphsLib.util import clear_data
 
 __all__ = [
-    'to_ufos', 'clear_data', 'set_redundant_data', 'set_custom_params',
+    'to_ufos', 'set_redundant_data', 'set_custom_params',
     'GLYPHS_PREFIX'
 ]
 
@@ -163,29 +164,6 @@ def to_ufos(data, include_instances=False, family_name=None, debug=False):
     elif include_instances:
         return result, instances
     return result
-
-
-def clear_data(data):
-    """Clear empty list or dict attributes in data.
-
-    This is used to determine what input data provided to to_ufos was not
-    loaded into an UFO."""
-
-    if isinstance(data, dict):
-        for key, val in data.items():
-            if not clear_data(val):
-                del data[key]
-        return data
-    elif isinstance(data, list):
-        i = 0
-        while i < len(data):
-            val = data[i]
-            if not clear_data(val):
-                del data[i]
-            else:
-                i += 1
-        return data
-    return True
 
 
 def generate_base_fonts(data, family_name):
